requirements.txt
Django==3.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-filter==2.2.0
django-nested-admin==3.3.3
django-phone-field==1.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0

